Add new module type_checker.rs
